# hashing can maybe be faster