@kescobo 2025-06-24
I was going to say that hashing kmers is pretty critical to do fast, but maybe we should leave that to `Kmers.jl` and expect that anything here would be longer?