@adienes 2025-06-24
it also probably wouldn't be that hard to fiddle with it and add a fast path here for short sequences. it would just be a bit less elegant since it's not a direct forward to `hash_shaped`