@adienes 2025-06-24
ok, makes sense. possibly parts of the implementation (the unrolling) are stealable, as you mention though 🙂